+++ cpp/lib_calvin/container/set_test.h
@@ -52,7 +52,7 @@ namespace lib_calvin_container
 	template <typename Impl>
 	void setRvalueTest(Impl &&impl, std::string);
 
-	void randomAccessSpeedTest(size_t size);
+	void randomAccessSpeedTest(int size);
 }
 
 #endif
